Moved get_data_config to test_data
